transferi metodo para app
